Merge #wp-content and #description IDs
Traditionally, "wp-content" has always been the ID of the element that's
wrapped around the content managed by WordPress's editor. While building
the wireframes, however, I used "description" to refer to the block that
contained the product description. This block ended up becoming the body
text's element, which is now causing confusion with #wp-content on other
pages (which I used for templates that lacked what a normal person would
consider a "description field"). Because I foresee this leading to messy
code and confusion later on, I'm merging the two selectors to spare some
unlucky coder a crapload of future headaches. Which will probably be me.