Remove black stripe from homepage's masthead
This was a relic from an earlier stage of the design phase, when the nav
links were originally located in the centre of the masthead, sandwiching
the DON® logo. The original idea was that they'd stay locked in position
the same as the logo, sticking onto the blue navbar as the user scrolled
downwards.

This was dropped in a creative revision later on, as it was decided that
having the menus placed in the centre of the masthead gave no indication
to users that there was content further down the page. The menu was then
moved to the bottom of the page so users had incentive to explore.

Somehow, the black stripe from the earlier designs survived the creative
process, and made it into the final design, despite serving no purpose.
It was unanimously agreed that the odd artefact should be removed.